+++ mlpf/model/plots.py
@@ -34,8 +34,8 @@ def log_oc_visualizations_to_tensorboard(batch, ypred_raw, ytarget, tensorboard_
     if mask.sum() == 0:
         return
 
-    X = batch.X[event_idx][mask].cpu().numpy()
-    ytarget_pn = ytarget["particle_number"][event_idx][mask].cpu().numpy()
+    X = batch.X[event_idx][mask].cpu().float().numpy()
+    ytarget_pn = ytarget["particle_number"][event_idx][mask].cpu().float().numpy()
 
     # Debug: Check if all PN are 0
     num_nonzero = np.sum(ytarget_pn > 0)
@@ -44,8 +44,8 @@ def log_oc_visualizations_to_tensorboard(batch, ypred_raw, ytarget, tensorboard_
     if len(ypred_raw) <= 5:
         return
 
-    oc_beta = ypred_raw[4][event_idx][mask].detach().cpu()
-    oc_coords = ypred_raw[5][event_idx][mask].detach().cpu()
+    oc_beta = ypred_raw[4][event_idx][mask].detach().cpu().float()
+    oc_coords = ypred_raw[5][event_idx][mask].detach().cpu().float()
 
     pred_pn = get_clustering(oc_beta, oc_coords).cpu().numpy()
 
@@ -137,17 +137,17 @@ def log_oc_visualizations_to_tensorboard(batch, ypred_raw, ytarget, tensorboard_
 
 
 def validation_plots(batch, ypred_raw, ytarget, ypred, tensorboard_writer, epoch, outdir):
-    X = batch.X[batch.mask].cpu()
-    ytarget_flat = batch.ytarget[batch.mask].cpu()
-    ypred_binary = ypred_raw[0][batch.mask].detach().cpu()
+    X = batch.X[batch.mask].cpu().float()
+    ytarget_flat = batch.ytarget[batch.mask].cpu().float()
+    ypred_binary = ypred_raw[0][batch.mask].detach().cpu().float()
     ypred_binary_cls = torch.argmax(ypred_binary, axis=-1)
-    ypred_cls = ypred_raw[1][batch.mask].detach().cpu()
-    ypred_p4 = ypred_raw[2][batch.mask].detach().cpu()
+    ypred_cls = ypred_raw[1][batch.mask].detach().cpu().float()
+    ypred_p4 = ypred_raw[2][batch.mask].detach().cpu().float()
 
     to_concat = [X, ytarget_flat, ypred_binary, ypred_cls, ypred_p4]
     if len(ypred_raw) > 4:
-        ypred_oc_beta = ypred_raw[4][batch.mask].detach().cpu()
-        ypred_oc_coords = ypred_raw[5][batch.mask].detach().cpu()
+        ypred_oc_beta = ypred_raw[4][batch.mask].detach().cpu().float()
+        ypred_oc_coords = ypred_raw[5][batch.mask].detach().cpu().float()
         to_concat.append(ypred_oc_beta)
         to_concat.append(ypred_oc_coords)
 
@@ -265,43 +265,43 @@ def validation_plots(batch, ypred_raw, ytarget, ypred, tensorboard_writer, epoch
             _logger.info(f"plotted sig_proba_elemtype{int(xcls)}")
 
         try:
-            tensorboard_writer.add_histogram("pt_target", torch.clamp(batch.ytarget[batch.mask][:, 2], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("pt_target", torch.clamp(batch.ytarget[batch.mask][:, 2].float(), -10, 10), global_step=epoch)
             _logger.info("plotted pt_target histogram")
-            tensorboard_writer.add_histogram("pt_pred", torch.clamp(ypred_raw[2][batch.mask][:, 0], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("pt_pred", torch.clamp(ypred_raw[2][batch.mask][:, 0].float(), -10, 10), global_step=epoch)
             _logger.info("plotted pt_pred histogram")
-            ratio = (ypred_raw[2][batch.mask][:, 0] / batch.ytarget[batch.mask][:, 2])[batch.ytarget[batch.mask][:, 0] != 0]
+            ratio = (ypred_raw[2][batch.mask][:, 0].float() / batch.ytarget[batch.mask][:, 2].float())[batch.ytarget[batch.mask][:, 0] != 0]
             tensorboard_writer.add_histogram("pt_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
             _logger.info("plotted pt_ratio histogram")
 
-            tensorboard_writer.add_histogram("eta_target", torch.clamp(batch.ytarget[batch.mask][:, 3], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("eta_target", torch.clamp(batch.ytarget[batch.mask][:, 3].float(), -10, 10), global_step=epoch)
             _logger.info("plotted eta_target histogram")
-            tensorboard_writer.add_histogram("eta_pred", torch.clamp(ypred_raw[2][batch.mask][:, 1], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("eta_pred", torch.clamp(ypred_raw[2][batch.mask][:, 1].float(), -10, 10), global_step=epoch)
             _logger.info("plotted eta_pred histogram")
-            ratio = (ypred_raw[2][batch.mask][:, 1] / batch.ytarget[batch.mask][:, 3])[batch.ytarget[batch.mask][:, 0] != 0]
+            ratio = (ypred_raw[2][batch.mask][:, 1].float() / batch.ytarget[batch.mask][:, 3].float())[batch.ytarget[batch.mask][:, 0] != 0]
             tensorboard_writer.add_histogram("eta_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
             _logger.info("plotted eta_ratio histogram")
 
-            tensorboard_writer.add_histogram("sphi_target", torch.clamp(batch.ytarget[batch.mask][:, 4], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("sphi_target", torch.clamp(batch.ytarget[batch.mask][:, 4].float(), -10, 10), global_step=epoch)
             _logger.info("plotted sphi_target histogram")
-            tensorboard_writer.add_histogram("sphi_pred", torch.clamp(ypred_raw[2][batch.mask][:, 2], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("sphi_pred", torch.clamp(ypred_raw[2][batch.mask][:, 2].float(), -10, 10), global_step=epoch)
             _logger.info("plotted sphi_pred histogram")
-            ratio = (ypred_raw[2][batch.mask][:, 2] / batch.ytarget[batch.mask][:, 4])[batch.ytarget[batch.mask][:, 0] != 0]
+            ratio = (ypred_raw[2][batch.mask][:, 2].float() / batch.ytarget[batch.mask][:, 4].float())[batch.ytarget[batch.mask][:, 0] != 0]
             tensorboard_writer.add_histogram("sphi_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
             _logger.info("plotted sphi_ratio histogram")
 
-            tensorboard_writer.add_histogram("cphi_target", torch.clamp(batch.ytarget[batch.mask][:, 5], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("cphi_target", torch.clamp(batch.ytarget[batch.mask][:, 5].float(), -10, 10), global_step=epoch)
             _logger.info("plotted cphi_target histogram")
-            tensorboard_writer.add_histogram("cphi_pred", torch.clamp(ypred_raw[2][batch.mask][:, 3], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("cphi_pred", torch.clamp(ypred_raw[2][batch.mask][:, 3].float(), -10, 10), global_step=epoch)
             _logger.info("plotted cphi_pred histogram")
-            ratio = (ypred_raw[2][batch.mask][:, 3] / batch.ytarget[batch.mask][:, 5])[batch.ytarget[batch.mask][:, 0] != 0]
+            ratio = (ypred_raw[2][batch.mask][:, 3].float() / batch.ytarget[batch.mask][:, 5].float())[batch.ytarget[batch.mask][:, 0] != 0]
             tensorboard_writer.add_histogram("cphi_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
             _logger.info("plotted cphi_ratio histogram")
 
-            tensorboard_writer.add_histogram("energy_target", torch.clamp(batch.ytarget[batch.mask][:, 6], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("energy_target", torch.clamp(batch.ytarget[batch.mask][:, 6].float(), -10, 10), global_step=epoch)
             _logger.info("plotted energy_target histogram")
-            tensorboard_writer.add_histogram("energy_pred", torch.clamp(ypred_raw[2][batch.mask][:, 4], -10, 10), global_step=epoch)
+            tensorboard_writer.add_histogram("energy_pred", torch.clamp(ypred_raw[2][batch.mask][:, 4].float(), -10, 10), global_step=epoch)
             _logger.info("plotted energy_pred histogram")
-            ratio = (ypred_raw[2][batch.mask][:, 4] / batch.ytarget[batch.mask][:, 6])[batch.ytarget[batch.mask][:, 0] != 0]
+            ratio = (ypred_raw[2][batch.mask][:, 4].float() / batch.ytarget[batch.mask][:, 6].float())[batch.ytarget[batch.mask][:, 0] != 0]
             tensorboard_writer.add_histogram("energy_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
             _logger.info("plotted energy_ratio histogram")
         except ValueError as e:
